+++ app/Console/Commands/FarmFunatic.php
@@ -106,7 +106,8 @@ class FarmFunatic extends Command
             } catch (\Throwable $e) {
                 /** Log Error */
                 Log::error('Funatic Error', [
-                    'message' => $e->getMessage()
+                    'message' => $e->getMessage(),
+                    'line' => $e->getLine()
                 ]);
             }
         })->filter();
@@ -214,11 +215,14 @@ class FarmFunatic extends Command
                         );
                     }
                 } catch (\Throwable $e) {
-                    $account->delete();
+                    if (app()->isProduction()) {
+                        $account->delete();
+                    }
 
                     /** Log Error */
                     Log::error('Funatic Error', [
-                        'message' => $e->getMessage()
+                        'message' => $e->getMessage(),
+                        'line' => $e->getLine()
                     ]);
                 }
             })->filter();
